Uptick dependencies
