+++ build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'org.jetbrains.kotlin.jvm' version '1.4.0'
+    id 'org.jetbrains.kotlin.jvm' version '1.6.0'
     id "org.jlleitschuh.gradle.ktlint" version "9.4.1"
 
 }
@@ -8,7 +8,6 @@ group 'com.lwise'
 version '0.0.1'
 
 repositories {
-    jcenter()
     mavenCentral()
     maven {
         url 'https://m2.dv8tion.net/releases'
@@ -51,22 +50,22 @@ configurations {
 dependencies {
     implementation "com.oracle.ojdbc:ojdbc8:19.3.0.0"
     implementation "io.github.cdimascio:dotenv-kotlin:6.2.2"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.4.0"
-    implementation "com.discord4j:discord4j-core:3.1.1"
-    implementation "ch.qos.logback:logback-classic:1.2.3"
-    implementation 'org.postgresql:postgresql:42.2.10'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.6.0'
+    implementation 'com.discord4j:discord4j-core:3.2.1'
+    implementation 'ch.qos.logback:logback-classic:1.2.7'
+    implementation 'org.postgresql:postgresql:42.3.1'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2-native-mt'
-    implementation 'com.squareup.okhttp3:okhttp:4.9.0'
+    implementation 'com.squareup.okhttp3:okhttp:4.9.2'
     implementation 'com.sedmelluq:lavaplayer:1.3.78'
     implementation 'se.michaelthelin.spotify:spotify-web-api-java:7.0.0'
-    compile 'org.json:json:20190722'
+    compile 'org.json:json:20210307'
 
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.3.1'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testImplementation 'org.hamcrest:hamcrest:2.2'
     testImplementation 'org.hamcrest:hamcrest-library:2.2'
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.4.2")
+    testRuntimeOnly('org.junit.jupiter:junit-jupiter-engine:5.8.1')
 
-    integrationTestCompile group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.7.0'
+    integrationTestCompile group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.8.1'
 
 }
