+++ src/main/kotlin/com/lwise/listeners/messages/AlignmentOptInListener.kt
@@ -30,7 +30,7 @@ class AlignmentOptInListener : MessageListener {
             }.filter { roleName ->
                 ConfigUtil.alignmentRoles.keys.contains(roleName)
             }.collectList()
-        val filteredRoles = userAlignmentRoles.block()!!
+        val filteredRoles = userAlignmentRoles.block()
         return if (!filteredRoles.isNullOrEmpty()) {
             // the user is already playing
             responseVector.channel.createMessage(getFailureResponseMessage())
@@ -44,7 +44,7 @@ class AlignmentOptInListener : MessageListener {
             DatabaseClient.update(userQuery)
 
             // give the user the True Neutral role to start
-            userToOptIn.addRole(Snowflake.of(ConfigUtil.alignmentRoles["True Neutral"])).block()
+            userToOptIn.addRole(Snowflake.of(ConfigUtil.alignmentRoles["True Neutral"]!!)).block()
             super.respond(responseVector)
         }
     }
