+++ src/main/kotlin/com/lwise/listeners/messages/musicplayer/PauseMusicListener.kt
@@ -1,18 +1,19 @@
-package com.lwise.listeners.messages
+package com.lwise.listeners.messages.musicplayer
 
+import com.lwise.listeners.messages.MessageListener
 import com.lwise.types.events.MessageEvent
 import com.lwise.util.player.MusicPlayer
 import discord4j.core.`object`.entity.Message
 import reactor.core.publisher.Mono
 
-class SkipSongListener : MessageListener {
+class PauseMusicListener : MessageListener {
     override val regexString: String
-        get() = "m!skip"
+        get() = "m!pause"
 
     override fun getResponseMessage(): String = ""
 
     override fun respond(responseVector: MessageEvent): Mono<Message> {
-        MusicPlayer.skip()
+        MusicPlayer.pause()
         return super.respond(responseVector)
     }
 }
