+++ src/main/kotlin/com/lwise/listeners/messages/musicplayer/RemoveFromQueueListener.kt
@@ -1,5 +1,6 @@
-package com.lwise.listeners.messages
+package com.lwise.listeners.messages.musicplayer
 
+import com.lwise.listeners.messages.MessageListener
 import com.lwise.types.events.MessageEvent
 import com.lwise.util.player.MusicPlayer
 import discord4j.core.`object`.entity.Message
