+++ src/main/kotlin/com/lwise/listeners/messages/musicplayer/VoiceJoinListener.kt
@@ -1,8 +1,10 @@
-package com.lwise.listeners.messages
+package com.lwise.listeners.messages.musicplayer
 
+import com.lwise.listeners.messages.MessageListener
 import com.lwise.types.events.MessageEvent
 import com.lwise.util.player.MusicPlayer
 import discord4j.core.`object`.entity.Message
+import discord4j.core.spec.VoiceChannelJoinSpec
 import reactor.core.publisher.Mono
 
 class VoiceJoinListener : MessageListener {
@@ -13,11 +15,12 @@ class VoiceJoinListener : MessageListener {
 
     override fun respond(responseVector: MessageEvent): Mono<Message> {
         val requestingUser = responseVector.author
+        val voiceJoinSpec = VoiceChannelJoinSpec.builder().apply {
+            provider(MusicPlayer.audioProvider)
+        }.build()
         requestingUser?.let { user ->
             user.voiceState.block()?.let { voiceState ->
-                voiceState.channel.block()?.join { spec ->
-                    spec.setProvider(MusicPlayer.audioProvider)
-                }?.block()
+                voiceState.channel.block()?.join(voiceJoinSpec)?.block()
             }
         }
         return Mono.empty()
